refactor: share chat agent helper
