+++ coderd/workspaceupdates.go
@@ -15,6 +15,7 @@ import (
 	"github.com/coder/coder/v2/coderd/rbac"
 	"github.com/coder/coder/v2/coderd/util/slice"
 	"github.com/coder/coder/v2/coderd/wspubsub"
+	"github.com/coder/coder/v2/coderd/x/chatd/chatagent"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/tailnet"
 	"github.com/coder/coder/v2/tailnet/proto"
@@ -280,7 +281,7 @@ func convertRows(rows []database.GetWorkspacesAndAgentsByOwnerIDRow) workspacesB
 	for _, row := range rows {
 		agents := []database.AgentIDNamePair{}
 		for _, agent := range row.Agents {
-			if isChatAgent(agent.Name) {
+			if chatagent.IsChatAgent(agent.Name) {
 				continue
 			}
 			agents = append(agents, database.AgentIDNamePair{
